Started using more components instead of a big block of code
