+++ client_side/pages/index.tsx
@@ -5,6 +5,7 @@ import Papa from 'papaparse';
 import HistoryButton from "../components/HistoryButton";
 import Footer from "../components/Footer";
 
+
 function FileUploadForm() {
     const {filename, setFilename} = useContext(UploadContext);
     const fileInput = useRef<HTMLInputElement>(null);
